+++ src/client/tcp_connection.cpp
@@ -6,6 +6,7 @@
 #include "bc/soup/logical_packets.h"
 #include "bc/soup/rw_packets.h"
 
+#include <cstdint>
 #include <utility>
 
 namespace bc::soup::client {
@@ -292,7 +293,8 @@ Write_error Tcp_connection::send_debug_packet(std::string_view text) {
   if (state_.state() == State::connecting || state_.is_closing())
     return Write_error::disconnected;
   return socket_.async_write(
-      Write_packet(Debug_packet::packet_type, text.data(), text.size()));
+      Write_packet(Debug_packet::packet_type, text.data(),
+                   static_cast<std::uint16_t>(text.size())));
 }
 
 void Tcp_connection::close() {
